refactor: Extract event bus subscribers into dedicated files
Move inline event bus subscribers from lifecycle-service.ts into
self-contained subscriber classes in src/subscribers/:

- ContextTrackingSubscriber: beginTurn, updateUsageMetadata, reset
  on session changes (4 subscriptions)
- AccessedFilesSubscriber: file access tracking on toolChainComplete
- Move ToolExecutionLogger from src/services/ to src/subscribers/

lifecycle-service.ts drops ~45 lines of inline handler logic and
becomes purely about service construction again. Each subscriber
follows the established pattern: constructor registers, stores
unsubscribers, destroy() cleans up.
